Add Toast Provider
