+++ ui/src/views/renders/[id]/RenderSidebar/CheckpointLimitEditor/CheckpointLimitForm.tsx
@@ -1,18 +1,9 @@
-import { HiExclamation } from 'react-icons/hi';
-import { useState } from 'react';
-
-import { Button, Spinner, Toast, ToastToggle, type ToastTheme } from 'flowbite-react';
 import { useSelector } from '@tanstack/react-store';
+import { Button, Spinner } from 'flowbite-react';
 import { useAppForm } from '@/hooks/useAppForm';
 import { useUpdateRenderTotalCheckpoints } from '@/hooks/useRenderMutations';
+import { useToast } from '@/providers/ToastProvider';
 import { z } from 'zod';
-import { motion, AnimatePresence } from 'framer-motion';
-import type { DeepPartial } from 'flowbite-react/types';
-
-type ToastItem = {
-  id: string;
-  message: string;
-};
 
 export type CheckpointLimitFormProps = {
   currentValue: number;
@@ -24,16 +15,7 @@ export function CheckpointLimitForm(props: CheckpointLimitFormProps) {
   const { currentValue, currentCheckpointIteration, renderID } = props;
 
   const { mutate: updateCheckpoints, isPending } = useUpdateRenderTotalCheckpoints();
-  const [toasts, setToasts] = useState<ToastItem[]>([]);
-
-  function createToast(message: string) {
-    const id = crypto.randomUUID();
-    setToasts((prev) => [...prev, { id, message }]);
-  }
-
-  function dismissToast(id: string) {
-    setToasts((prev) => prev.filter((t) => t.id !== id));
-  }
+  const { createToast } = useToast();
 
   const form = useAppForm({
     defaultValues: { total_checkpoints: currentValue },
@@ -64,12 +46,6 @@ export function CheckpointLimitForm(props: CheckpointLimitFormProps) {
     );
   }
 
-  const toastTheme: DeepPartial<ToastTheme> = {
-    root: {
-      base: 'max-w-lg',
-    },
-  };
-
   return (
     <>
       <form.AppField name="total_checkpoints">
@@ -93,31 +69,6 @@ export function CheckpointLimitForm(props: CheckpointLimitFormProps) {
           'Update'
         )}
       </Button>
-
-      <div className="fixed right-4 bottom-4 z-50 flex flex-col gap-2">
-        <AnimatePresence>
-          {toasts.map((toast) => (
-            <motion.div
-              key={toast.id}
-              className="max-w-lg"
-              initial={{ x: '100%', opacity: 0 }}
-              animate={{ x: 0, opacity: 1 }}
-              exit={{ x: '100%', opacity: 0 }}
-              transition={{ type: 'spring', stiffness: 300, damping: 30 }}
-            >
-              <Toast theme={toastTheme}>
-                <div className="inline-flex h-8 w-8 shrink-0 items-center justify-center rounded-lg bg-orange-100 text-orange-500 dark:bg-orange-700 dark:text-orange-200">
-                  <HiExclamation className="h-5 w-5" />
-                </div>
-                <div className="ml-3 text-sm font-normal text-red-600 dark:text-red-400">
-                  {toast.message}
-                </div>
-                <ToastToggle onDismiss={() => dismissToast(toast.id)} />
-              </Toast>
-            </motion.div>
-          ))}
-        </AnimatePresence>
-      </div>
     </>
   );
 }
